requirements.txt
jsonpickle
scikit-learn
pandas
pydub
numpy
mysql-connector-python
librosa